Add files via upload
This code is a JavaScript script that retrieves and displays current weather information for a user-specified location using the OpenWeatherMap API.

The code starts by getting references to three DOM elements in the HTML document: a form element with the ID "location-form", a text input element with the ID "location", and a div element with the ID "weather-info".

Next, the script sets the API key and base URL for the OpenWeatherMap API.

Then, an event listener is added to the form submission, which prevents the default behavior of the page reloading upon form submission. The script retrieves the location entered by the user, sends an API request to OpenWeatherMap, and parses the response data.

If the API request is successful, the script updates the UI with the retrieved weather information by populating the "weather-info" div element with the location name, temperature, and description. If there is an error with the API request, the script logs an error message to the console and updates the "weather-info" div element to display an error message.